Changed color calculation. Dirty hack, needs a settings menu option.
Color is now determined by assigning the time of day to the hue part of a HSV value. Since there's 1440 minutes in a day, and there's 360 degrees on a HSV hue circle, each hue is shown for 4 minutes.

Downside: The numbers shown are not the color shown. But, at least this way we get to see all colors instead of mostly black, blues and greens.

NOTE: It helps if I actually commit the change.